+++ test/blackbox-tests/test-cases/melange/promote-with-lib.t
@@ -8,7 +8,7 @@ Test melange.emit promotion
   $ mkdir lib
   $ cat > lib/dune <<EOF
   > (library
-  >  (modes melange)
+  >  (modes :standard melange)
   >  (name mylib))
   > EOF
 
@@ -32,7 +32,29 @@ Test melange.emit promotion
   >   print_endline "hello"
   > EOF
 
+Fails with an informative error message if we parsed OSL for modes
+
   $ dune build @dist
+  File "lib/dune", line 1, characters 0-50:
+  1 | (library
+  2 |  (modes :standard melange)
+  3 |  (name mylib))
+  Error: Ordered set language for modes is only available since version 3.8 of
+  the dune language. Please update your dune-project file to have (lang dune
+  3.8).
+  [1]
+
+
+  $ cat > dune-project <<EOF
+  > (lang dune 3.8)
+  > (using melange 0.1)
+  > EOF
+  $ dune build @dist
+
+Library has `(modes :standard)` so it also builds for bytecode / native
+
+  $ dune build lib/.mylib.objs/byte/mylib.cmo
+  $ dune build lib/.mylib.objs/native/mylib.cmx
 
 Targets are promoted to the source tree